requirement.txt
streamlit
torch  # this will usually work if CPU wheels are available
torchvision
Pillow
requests
